Rearrange Kalibro Configurations' tests
 * Now each class has its own test file

Signed off by: Diego Araújo <diegoamc90@gmail.com>
